Utilize measureText to determine title length.
While measureText is used frequently throughout LiteGraph, computeSize
calls did not make use of it. As a result, the node size calculations
are consistently incorrect for long titles, short titles, or titles that
utilize characters of abnormal width such as emojis.

My assumption is that this was purely because a context with the
required title font isn't available when the call is made, but it's easy
to add one just for ensuring the accuracy of text sizing.