Add fallback for virtual nodes that redirect links
KJNodes provides Get and Set nodes that overload getInputLink to
to designate the calculated link source.

If the existing check (which simply returns a corresponding input link
by slotid) fails, a call to node.getInputLink is attempted.
If there are other virtual nodes that were broken by the subgraph
implementation (such a switch that has multiple input links, but
resolves to only one), this fallback may need to instead replace the
existing check.