Fix inconsistent subgraphNode usage
Prior to this commit, subgraphNode inconsistently refers to either the
parent graph, or to indicate the current node is a subgraph.

This corrects the usage of subgraphNode to consistently refer to the
subgraph instance as defined in the constructor.

This requires a corresponding frontend commit to function, and solves a
bug where graph serialization fails due to an incorrectly reported
infinite loop.